requirements.txt
httpx>=0.27.0
pyyaml>=6.0
web3>=6.20.0
eth-account>=0.13.0
fastapi>=0.115.0
uvicorn>=0.30.0
